refactor: Change `order class` package in domain
