+++ src/main/java/milkcoke/core/service/order/OrderService.java
@@ -1,6 +1,6 @@
 package milkcoke.core.service.order;
 
-import milkcoke.core.order.Order;
+import milkcoke.core.domain.order.Order;
 
 public interface OrderService {
     Order createOrder(Long memberId, String itemName, int itemPrice);
